+++ packages/node/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Sync with node-core, fixed modulo block ahead of finalized block issue (#2132)
 ### Changed
 - Use WorkerInMemoryCacheService from node core (#2125)